Implemented an overriddalble image mapper for use in product variant mapping
