+++ src/Enterspeed.Commercetools.Integration/Domain/Mappers/EnterspeedVariantMapper.cs
@@ -15,6 +15,7 @@ public class EnterspeedVariantMapper : IMapper<ProductVariantMappingContext, Ent
     private readonly IMapper<List<IAsset>, List<IEnterspeedProperty>> _assetMapper;
     private readonly IMapper<IProductVariantAvailability, IEnterspeedProperty> _availabilityMapper;
     private readonly IMapper<List<IPrice>, IEnterspeedProperty> _priceMapper;
+    private readonly IMapper<List<IImage>, IEnterspeedProperty> _imageMapper;
     private readonly IProductVariantIdFactory _productVariantIdFactory;
     private readonly IEnterspeedEntityTypeProvider _entityTypeProvider;
 
@@ -23,12 +24,14 @@ public class EnterspeedVariantMapper : IMapper<ProductVariantMappingContext, Ent
         IMapper<List<IAsset>, List<IEnterspeedProperty>> assetMapper,
         IMapper<IProductVariantAvailability, IEnterspeedProperty> availabilityMapper,
         IMapper<List<IPrice>, IEnterspeedProperty> priceMapper,
+        IMapper<List<IImage>, IEnterspeedProperty> imageMapper,
         IProductVariantIdFactory productVariantIdFactory,
         IEnterspeedEntityTypeProvider entityTypeProvider)
     {
         _propertyMapper = propertyMapper;
         _assetMapper = assetMapper;
         _priceMapper = priceMapper;
+        _imageMapper = imageMapper;
         _productVariantIdFactory = productVariantIdFactory;
         _entityTypeProvider = entityTypeProvider;
         _availabilityMapper = availabilityMapper;
@@ -41,16 +44,17 @@ public class EnterspeedVariantMapper : IMapper<ProductVariantMappingContext, Ent
         var propertiesTask = _propertyMapper.MapAsync(source.Attributes);
         var assetTask = _assetMapper.MapAsync(source.Assets);
         var priceTask = _priceMapper.MapAsync(source.Prices);
+        var imageTask = _imageMapper.MapAsync(source.Images);
         var variantIdTask = _productVariantIdFactory.GetProductVariantIdAsync(context.Product, source);
         var typeTask = _entityTypeProvider.GetEntityTypeAsync(source);
 
-        Task.WaitAll(propertiesTask, assetTask, priceTask, variantIdTask, typeTask);
+        Task.WaitAll(propertiesTask, assetTask, priceTask, imageTask, variantIdTask, typeTask);
 
         var properties = new Dictionary<string, IEnterspeedProperty>()
         {
             ["id"] = new StringEnterspeedProperty(await variantIdTask),
             ["prices"] = await priceTask,
-            ["images"] = MapImage(source.Images),
+            ["images"] = await imageTask,
             ["attributes"] = await propertiesTask,
             ["assets"] = new ArrayEnterspeedProperty(string.Empty, (await assetTask).ToArray())
         };
@@ -75,21 +79,5 @@ public class EnterspeedVariantMapper : IMapper<ProductVariantMappingContext, Ent
             Properties = properties,
             ParentId = context.Product.Id
         };
-    }
-
-    private static IEnterspeedProperty MapImage(List<IImage> source)
-    {
-        var images = source.Select(image => new ObjectEnterspeedProperty(new Dictionary<string, IEnterspeedProperty>
-        {
-            ["url"] = new StringEnterspeedProperty(image.Url),
-            ["label"] = new StringEnterspeedProperty(image.Label),
-            ["dimensions"] = new ObjectEnterspeedProperty(new Dictionary<string, IEnterspeedProperty>
-            {
-                ["w"] = new NumberEnterspeedProperty(image.Dimensions.W),
-                ["h"] = new NumberEnterspeedProperty(image.Dimensions.H)
-            })
-        })).ToArray();
-
-        return new ArrayEnterspeedProperty(string.Empty, images);
     }
 }
